Created UITabBarController Class and made changes based on setting up in own class
